fix: fix issue with circular deps
