+++ packages/vrender-core/src/common/segment/curve/cubic-bezier.ts
@@ -35,35 +35,6 @@ export function divideCubic(curve: ICubicBezierCurve, t: number): ICubicBezierCu
 
   return [curve1, curve2];
 }
-/**
- * 对三次贝塞尔曲线进行分割
- * @param p0 起点
- * @param p1 控制点1
- * @param p2 控制点2
- * @param p3 终点
- * @param t
- */
-export function divideQuad(curve: IQuadraticBezierCurve, t: number): IQuadraticBezierCurve[] {
-  const { p0, p1, p2 } = curve;
-
-  // 划分点
-  const pt = quadPointAt(p0, p1, p2, t);
-  //  const xt = pt.x;
-  //  const yt = pt.y;
-
-  // 计算两点之间的差值点
-  const c1 = PointService.pointAtPP(p0, p1, t);
-  const c2 = PointService.pointAtPP(p1, p2, t);
-  // const c3 = PointService.pointAtPP(p2, p3, t);
-  // const c12 = PointService.pointAtPP(c1, c2, t);
-  // const c23 = PointService.pointAtPP(c2, c3, t);
-  // const direction = p1.x1 ? p1.y > p0.y ? 0 : 1 : p1.x > p0.x ? 0 : 1;
-
-  const curve1 = new QuadraticBezierCurve(p0, c1, pt);
-  const curve2 = new QuadraticBezierCurve(pt, c2, p2);
-
-  return [curve1, curve2];
-}
 
 export class CubicBezierCurve extends Curve implements ICubicBezierCurve {
   type: number = CurveTypeEnum.CubicBezierCurve;
